backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0
yfinance==0.2.31
httpx==0.25.0
aiofiles==23.2.1
python-multipart==0.0.6
pytest==7.4.2
pytest-asyncio==0.21.1
jupyter==1.0.0
matplotlib==3.8.0 